requirements.txt
Django>=4.2.0
instaloader>=4.10.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0 